12.04: избавился от лишнего и настроил редирект из списка избранного в сам пост

+++ feedback-service/src/main/java/org/alexreverse/repository/FavouritePostRepository.java
@@ -6,7 +6,7 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 
-public interface FavouritePostRepository extends ReactiveCrudRepository<FavouritePost, Integer> {
+public interface FavouritePostRepository extends ReactiveCrudRepository<FavouritePost, Long> {
 
     Mono<Void> deleteByPostIdAndUserName(int postId, String userName);
